+++ docs/components/index-page/index.tsx
@@ -19,7 +19,7 @@ function I18n() {
   const [active, setActive] = useState('')
 
   return (
-    <div className={cn(styles.comparison)}>
+    <div className={styles.comparison}>
       <div style={{ display: 'flex', flexDirection: 'column', gap: '.5rem' }}>
         {LANGUAGES.map(({ lang }) => (
           <span
@@ -64,7 +64,10 @@ export const IndexPage = () => (
         Simple, powerful and flexible site generation framework{' '}
         <br className="max-md:hidden" />
         with everything you love from{' '}
-        <Link href="https://nextjs.org">Next.js</Link>.
+        <Link href="https://nextjs.org" className="text-current">
+          Next.js
+        </Link>
+        .
       </p>
       <p className="subtitle">
         <Link className={styles.cta} href="/docs">