+++ docs/next.config.mjs
@@ -1,3 +1,4 @@
+import path from 'node:path'
 import nextra from 'nextra'
 
 const withNextra = nextra({
@@ -9,6 +10,10 @@ const withNextra = nextra({
   },
   defaultShowCopyCode: true
 })
+
+const sep = path.sep === '/' ? '/' : '\\\\'
+
+const ALLOWED_SVG_REGEX = new RegExp(`components${sep}icons${sep}.+\\.svg$`)
 
 export default withNextra({
   reactStrictMode: true,
@@ -29,15 +34,13 @@ export default withNextra({
     }
   ],
   webpack(config) {
-    const allowedSvgRegex = /components\/icons\/.+\.svg$/
-
     const fileLoaderRule = config.module.rules.find(rule =>
       rule.test?.test?.('.svg')
     )
-    fileLoaderRule.exclude = allowedSvgRegex
+    fileLoaderRule.exclude = ALLOWED_SVG_REGEX
 
     config.module.rules.push({
-      test: allowedSvgRegex,
+      test: ALLOWED_SVG_REGEX,
       use: ['@svgr/webpack']
     })
     return config