+++ packages/nextra/src/server/compile.ts
@@ -170,6 +170,7 @@ export async function compileMdx(
     }
   }
 
+  const fileCompatible = filePath ? { value: source, path: filePath } : source
   if (isPageMapImport) {
     const compiler = createProcessor({
       format,
@@ -178,9 +179,7 @@ export async function compileMdx(
         remarkMdxFrontMatter
       ]
     })
-    const vFile = await compiler.process(
-      filePath ? { value: source, path: filePath } : source
-    )
+    const vFile = await compiler.process(fileCompatible)
     const content = vFile.toString()
 
     const index = content.lastIndexOf('function _createMdxContent(props) {')
@@ -224,7 +223,6 @@ export async function compileMdx(
   const processor = compiler()
 
   try {
-    const fileCompatible = filePath ? { value: source, path: filePath } : source
     const vFile = await processor.process(fileCompatible)
 
     const data = vFile.data as {