+++ packages/nextra-theme-blog/CHANGELOG.md
@@ -1,5 +1,27 @@
 # nextra-theme-blog
 
+## 3.0.0-alpha.7
+
+### Patch Changes
+
+- 0b5cc9d5: make nextra compatible with windows
+- Updated dependencies [0b5cc9d5]
+  - nextra@3.0.0-alpha.7
+
+## 3.0.0-alpha.6
+
+### Patch Changes
+
+- Updated dependencies [03da778a]
+  - nextra@3.0.0-alpha.6
+
+## 3.0.0-alpha.5
+
+### Patch Changes
+
+- Updated dependencies [a3b67aea]
+  - nextra@3.0.0-alpha.5
+
 ## 3.0.0-alpha.4
 
 ### Patch Changes